server/requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_cors
flask_jwt_extended
psycopg2-binary
requests
gunicorn
python-dotenv
flask_mail
flask_restful
flask_marshmallow
marshmallow-sqlalchemy
flask_bcrypt
flask_login
flask_limiter
flask_jwt_extended
authlib